+++ public/assets/scripts/choices.search-prefix.js
@@ -2200,8 +2200,12 @@
             this._store.subscribe(this._render);
             this._render();
             this._addEventListeners();
-            var shouldDisable = !this.config.addItems ||
-                this.passedElement.element.hasAttribute('disabled');
+            var shouldDisable = (this._isTextElement && !this.config.addItems) ||
+                (this._isSelectElement &&
+                    !this.config.addChoices &&
+                    this._store.choices.length === 0) ||
+                this.passedElement.element.hasAttribute('disabled') ||
+                !!this.passedElement.element.closest('fieldset:disabled');
             if (shouldDisable) {
                 this.disable();
             }