+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bug Fixes (from 10.2.0)
 * Fix Choices does not accept an element from an iframe [#1057](https://github.com/Choices-js/Choices/issues/1057)
+* Fix Choices was not disable in a `<fieldset disabled>` [#1132](https://github.com/Choices-js/Choices/issues/1132)
 
 ## [11.0.0-rc5] (2024-08-08)